# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Youleap Dev Test — Product API",
    "description": "Mock product API for the Youleap mid-level developer recruitment test. Serves 12 products from static JSON with simulated latency.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev server"
    }
  ],
  "paths": {
    "/api/products": {
      "get": {
        "operationId": "listProducts",
        "summary": "List products",
        "description": "Returns a paginated, filterable list of products.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 12 },
            "description": "Items per page"
          },
          {
            "name": "offset",
            "in": "query",
            "schema": { "type": "integer", "default": 0 },
            "description": "Pagination offset"
          },
          {
            "name": "q",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Search by title or description (case-insensitive)"
          },
          {
            "name": "collection",
            "in": "query",
            "schema": { "type": "string", "enum": ["audio", "office", "lifestyle", "health", "home"] },
            "description": "Filter by collection handle"
          },
          {
            "name": "tag",
            "in": "query",
            "schema": { "type": "string", "enum": ["electronics", "audio", "furniture", "office", "accessories", "smart", "peripherals", "lighting", "bags", "nutrition", "fitness", "kitchen", "coffee", "home", "textiles", "yoga"] },
            "description": "Filter by tag value"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated product list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProductsResponse" }
              }
            }
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "operationId": "getProduct",
        "summary": "Get product by ID",
        "description": "Returns a single product by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Product ID (e.g. prod_01)",
            "examples": {
              "headphones": { "value": "prod_01" },
              "chair": { "value": "prod_02" },
              "bottle": { "value": "prod_03" }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Single product",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProductResponse" }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Product not found" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Price": {
        "type": "object",
        "required": ["amount", "currency_code"],
        "properties": {
          "amount": { "type": "integer", "description": "Price in agorot (cents). 29900 = ₪299.00", "example": 29900 },
          "currency_code": { "type": "string", "example": "ILS" }
        }
      },
      "ProductVariant": {
        "type": "object",
        "required": ["id", "title", "sku", "prices", "inventory_quantity"],
        "properties": {
          "id": { "type": "string", "example": "var_01a" },
          "title": { "type": "string", "example": "Black" },
          "sku": { "type": "string", "example": "HP-BLK-01" },
          "prices": { "type": "array", "items": { "$ref": "#/components/schemas/Price" } },
          "inventory_quantity": { "type": "integer", "description": "0 means out of stock", "example": 25 }
        }
      },
      "ProductOption": {
        "type": "object",
        "required": ["id", "title", "values"],
        "properties": {
          "id": { "type": "string", "example": "opt_color_01" },
          "title": { "type": "string", "example": "Color" },
          "values": { "type": "array", "items": { "type": "string" }, "example": ["Black", "White", "Navy Blue"] }
        }
      },
      "ProductImage": {
        "type": "object",
        "required": ["url"],
        "properties": {
          "url": { "type": "string", "format": "uri", "example": "https://picsum.photos/seed/headphones1/800/800" }
        }
      },
      "ProductTag": {
        "type": "object",
        "required": ["value"],
        "properties": {
          "value": { "type": "string", "example": "electronics" }
        }
      },
      "ProductCollection": {
        "type": "object",
        "required": ["id", "title", "handle"],
        "properties": {
          "id": { "type": "string", "example": "col_01" },
          "title": { "type": "string", "example": "Audio" },
          "handle": { "type": "string", "example": "audio" }
        }
      },
      "Product": {
        "type": "object",
        "required": ["id", "title", "handle", "description", "thumbnail", "images", "variants", "options", "tags", "collection", "created_at"],
        "properties": {
          "id": { "type": "string", "example": "prod_01" },
          "title": { "type": "string", "example": "Classic Wireless Headphones" },
          "handle": { "type": "string", "example": "classic-wireless-headphones" },
          "description": { "type": "string" },
          "thumbnail": { "type": "string", "format": "uri", "example": "https://picsum.photos/seed/headphones/400/400" },
          "images": { "type": "array", "items": { "$ref": "#/components/schemas/ProductImage" } },
          "variants": { "type": "array", "items": { "$ref": "#/components/schemas/ProductVariant" } },
          "options": { "type": "array", "items": { "$ref": "#/components/schemas/ProductOption" } },
          "tags": { "type": "array", "items": { "$ref": "#/components/schemas/ProductTag" } },
          "collection": { "$ref": "#/components/schemas/ProductCollection" },
          "created_at": { "type": "string", "format": "date-time", "example": "2026-01-15T10:00:00Z" }
        }
      },
      "ProductsResponse": {
        "type": "object",
        "required": ["products", "count", "limit", "offset"],
        "properties": {
          "products": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } },
          "count": { "type": "integer", "example": 12 },
          "limit": { "type": "integer", "example": 12 },
          "offset": { "type": "integer", "example": 0 }
        }
      },
      "ProductResponse": {
        "type": "object",
        "required": ["product"],
        "properties": {
          "product": { "$ref": "#/components/schemas/Product" }
        }
      }
    }
  }
}